renaming by jim's request: UrlParam => PageParam (type), createUrlParam => createPageParam (helper)
