+++ src/extensions/renderer-api/navigation.ts
@@ -1,3 +1,4 @@
-export { createUrlParam, navigate, isActiveRoute, UrlParamInit, UrlParam } from "../../renderer/navigation";
+export { PageParam, PageParamInit } from "../../renderer/navigation/page-param";
+export { navigate, isActiveRoute, createPageParam } from "../../renderer/navigation";
 export { hideDetails, showDetails, getDetailsUrl } from "../../renderer/components/kube-object/kube-object-details";
 export { IURLParams } from "../../common/utils/buildUrl";
